requirements.txt
peewee==3.14.0
Pillow==8.1.0
python-dateutil==2.8.1
six==1.15.0
